requirements.txt
pyserial
pandas
jupyter
matplotlib